+++ test/bigtest/interactors/filters/checkbox-filter.js
@@ -1,0 +1,7 @@
+import CheckboxInteractor from '@folio/stripes-smart-components/lib/SearchAndSort/components/CheckboxFilter/tests/interactor';
+
+import FilterInteractor from './filter';
+
+export default class CheckboxFilterInteractor extends FilterInteractor {
+  checkboxes = new CheckboxInteractor();
+}